fix(application.properties, MenuProduct, etc.): add an option "ddl-auto=validate" for JPA and fix the type of attribute quantity

+++ src/main/java/kitchenpos/domain/Product.java
@@ -44,7 +44,7 @@ public class Product {
         return unitPrice.getValue();
     }
 
-    public BigDecimal getPrice(int quantity) {
+    public BigDecimal getPrice(Long quantity) {
         return unitPrice.getPrice(quantity);
     }
 }
